move _combine_* functions to their own utils module
